Split out pyx to also pxd
This allows classes exposed to Python to be the same name as those in
the original c++ library.  Decided to use a single pxd and pyx file to
manage the library interaction, so that only one extension module needs
to be built and the setup.py is simpler.  This may need to change as
more of the GeoTess library is exposed.