Implement GeoTessGrid.getVertex using NumPy
Also, replace array.array usage in EarthShape.getVectorDegrees.  NumPy
just offers more options for dealing with C memory, and there is way
more online documentation for using NumPy arrays than array.array.

GeoTessGrid.getVertex returns a NumPy array that is still tied to the
in-memory grid, so changes there propagate to the wrapped C++ object.
Not so, for getVectorDegrees.