Add "from_geotessgrid" classmethod to Grid
This allows a PyGeoTess Grid object to be created from a
libgeotess.GeoTessGrid instance.  To help the constructor, I also allow
for an empty __init__ Grid constructor.  The .triangles and .vertices
methods will be moved to the Grid class from the Model class.  Those
methods will be re-directs to the attached Grid instance.