Allow cimport for pxd files
From 0.29 docx: One caveat if you use setuptools instead of distutils,
the default action when running python setup.py install is to create a
zipped egg file which will not work with cimport for pxd files when you
try to use them from a dependent package. To prevent this, include
zip_safe=False in the arguments to setup().
Hope this fixes things with the re-org branch.